Bugfix float to int conversion  - float point coordinates crashed cv2.line

+++ src/data_process/kitti_bev_utils.py
@@ -164,7 +164,7 @@ def drawRotatedBox(img, x, y, w, l, yaw, color):
     bev_corners = get_corners(x, y, w, l, yaw)
     corners_int = bev_corners.reshape(-1, 1, 2).astype(int)
     cv2.polylines(img, [corners_int], True, color, 2)
-    corners_int = bev_corners.reshape(-1, 2)
+    corners_int = bev_corners.reshape(-1, 2).astype(int)
     cv2.line(img, (corners_int[0, 0], corners_int[0, 1]), (corners_int[3, 0], corners_int[3, 1]), (255, 255, 0), 2)
